Password hash login (#13)
* Added password hashing

Now when registering, the password filled in the form is hashed and saved as a hash []byte. Package bcrypto was downloaded for this, therefore the changes in go.mod and go.sum.

* Login w password

Now the login requires the user to input the correct password before being able to login. Also updated .gitignore to ignore all files with the word test in it.